+++ gradio/components/code.py
@@ -2,13 +2,12 @@
 
 from __future__ import annotations
 
-import warnings
 from pathlib import Path
 from typing import Any, Literal
 
 from gradio_client.documentation import document, set_documentation_group
 
-from gradio.components.base import Component, _Keywords
+from gradio.components.base import Component
 from gradio.events import Events
 
 set_documentation_group("component")
@@ -57,6 +56,7 @@ class Code(Component):
         ]
         | None = None,
         *,
+        every: float | None = None,
         lines: int = 5,
         label: str | None = None,
         interactive: bool | None = None,
@@ -67,13 +67,16 @@ class Code(Component):
         visible: bool = True,
         elem_id: str | None = None,
         elem_classes: list[str] | str | None = None,
-        **kwargs,
+        render: bool = True,
+        root_url: str | None = None,
+        _skip_init_processing: bool = False,
     ):
         """
         Parameters:
             value: Default value to show in the code editor. If callable, the function will be called whenever the app loads to set the initial value of the component.
+            every: If `value` is a callable, run the function 'every' number of seconds while the client connection is open. Has no effect otherwise. Queue must be enabled. The event can be accessed (e.g. to cancel it) via this component's .load_event attribute.
             language: The language to display the code as. Supported languages listed in `gr.Code.languages`.
-            label: component name in interface.
+            label: The label for this component. Appears above the component and is also used as the header if there are a table of examples for this component. If None and used in a `gr.Interface`, the label will be the name of the parameter this component is assigned to.
             interactive: Whether user should be able to enter code or only view it.
             show_label: if True, will display label.
             container: If True, will place the component in a container - providing some extra padding around the border.
@@ -82,6 +85,8 @@ class Code(Component):
             visible: If False, component will be hidden.
             elem_id: An optional string that is assigned as the id of this component in the HTML DOM. Can be used for targeting CSS styles.
             elem_classes: An optional list of strings that are assigned as the classes of this component in the HTML DOM. Can be used for targeting CSS styles.
+            render: If False, component will not render be rendered in the Blocks context. Should be used if the intention is to assign event listeners now but render the component later.
+            root_url: The remote URL that of the Gradio app that this component belongs to. Used in `gr.load()`. Should not be set manually.
         """
         if language not in Code.languages:
             raise ValueError(f"Language {language} not supported.")
@@ -90,6 +95,7 @@ class Code(Component):
         self.lines = lines
         super().__init__(
             label=label,
+            every=every,
             interactive=interactive,
             show_label=show_label,
             container=container,
@@ -98,8 +104,10 @@ class Code(Component):
             visible=visible,
             elem_id=elem_id,
             elem_classes=elem_classes,
+            render=render,
+            root_url=root_url,
+            _skip_init_processing=_skip_init_processing,
             value=value,
-            **kwargs,
         )
 
     def preprocess(self, x: Any) -> Any:
@@ -122,47 +130,3 @@ class Code(Component):
 
     def example_inputs(self) -> Any:
         return "print('Hello World')"
-
-    @staticmethod
-    def update(
-        value: str
-        | tuple[str]
-        | None
-        | Literal[_Keywords.NO_VALUE] = _Keywords.NO_VALUE,
-        label: str | None = None,
-        show_label: bool | None = None,
-        container: bool | None = None,
-        scale: int | None = None,
-        min_width: int | None = None,
-        visible: bool | None = None,
-        language: Literal[
-            "python",
-            "markdown",
-            "json",
-            "html",
-            "css",
-            "javascript",
-            "typescript",
-            "yaml",
-            "dockerfile",
-            "shell",
-            "r",
-        ]
-        | None = None,
-        interactive: bool | None = None,
-    ):
-        warnings.warn(
-            "Using the update method is deprecated. Simply return a new object instead, e.g. `return gr.Code(...)` instead of `return gr.Code.update(...)`."
-        )
-        return {
-            "label": label,
-            "show_label": show_label,
-            "container": container,
-            "scale": scale,
-            "min_width": min_width,
-            "visible": visible,
-            "value": value,
-            "language": language,
-            "interactive": interactive,
-            "__type__": "update",
-        }