+++ gradio/components/plot.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 import json
-import warnings
 from types import ModuleType
 from typing import Any, Callable, Literal
 
@@ -12,9 +11,8 @@ import pandas as pd
 from gradio_client.documentation import document, set_documentation_group
 
 from gradio import processing_utils
-from gradio.components.base import Component, _Keywords
+from gradio.components.base import Component
 from gradio.data_classes import GradioModel
-from gradio.deprecation import warn_style_method_deprecation
 from gradio.events import Events
 
 set_documentation_group("component")
@@ -33,7 +31,7 @@ class AltairPlotData(PlotData):
 @document()
 class Plot(Component):
     """
-    Used to display various kinds of plots (matplotlib, plotly, or bokeh are supported)
+    Used to display various kinds of plots (matplotlib, plotly, or bokeh are supported).
     Preprocessing: this component does *not* accept input.
     Postprocessing: expects either a {matplotlib.figure.Figure}, a {plotly.graph_objects._figure.Figure}, or a {dict} corresponding to a bokeh plot (json_item format)
 
@@ -57,12 +55,14 @@ class Plot(Component):
         visible: bool = True,
         elem_id: str | None = None,
         elem_classes: list[str] | str | None = None,
-        **kwargs,
+        render: bool = True,
+        root_url: str | None = None,
+        _skip_init_processing: bool = False,
     ):
         """
         Parameters:
             value: Optionally, supply a default plot object to display, must be a matplotlib, plotly, altair, or bokeh figure, or a callable. If callable, the function will be called whenever the app loads to set the initial value of the component.
-            label: component name in interface.
+            label: The label for this component. Appears above the component and is also used as the header if there are a table of examples for this component. If None and used in a `gr.Interface`, the label will be the name of the parameter this component is assigned to.
             every: If `value` is a callable, run the function 'every' number of seconds while the client connection is open. Has no effect otherwise. Queue must be enabled. The event can be accessed (e.g. to cancel it) via this component's .load_event attribute.
             show_label: if True, will display label.
             container: If True, will place the component in a container - providing some extra padding around the border.
@@ -71,6 +71,8 @@ class Plot(Component):
             visible: If False, component will be hidden.
             elem_id: An optional string that is assigned as the id of this component in the HTML DOM. Can be used for targeting CSS styles.
             elem_classes: An optional list of strings that are assigned as the classes of this component in the HTML DOM. Can be used for targeting CSS styles.
+            render: If False, component will not render be rendered in the Blocks context. Should be used if the intention is to assign event listeners now but render the component later.
+            root_url: The remote URL that of the Gradio app that this component belongs to. Used in `gr.load()`. Should not be set manually.
         """
         super().__init__(
             label=label,
@@ -82,8 +84,10 @@ class Plot(Component):
             visible=visible,
             elem_id=elem_id,
             elem_classes=elem_classes,
+            render=render,
+            root_url=root_url,
+            _skip_init_processing=_skip_init_processing,
             value=value,
-            **kwargs,
         )
 
     def get_config(self):
@@ -97,31 +101,6 @@ class Plot(Component):
         config = super().get_config()
         config["bokeh_version"] = bokeh_version
         return config
-
-    @staticmethod
-    def update(
-        value: Any | Literal[_Keywords.NO_VALUE] | None = _Keywords.NO_VALUE,
-        label: str | None = None,
-        show_label: bool | None = None,
-        container: bool | None = None,
-        scale: int | None = None,
-        min_width: int | None = None,
-        visible: bool | None = None,
-    ):
-        warnings.warn(
-            "Using the update method is deprecated. Simply return a new object instead, e.g. `return gr.Plot(...)` instead of `return gr.Plot.update(...)`."
-        )
-        updated_config = {
-            "label": label,
-            "show_label": show_label,
-            "container": container,
-            "scale": scale,
-            "min_width": min_width,
-            "visible": visible,
-            "value": value,
-            "__type__": "update",
-        }
-        return updated_config
 
     def preprocess(self, x: Any) -> Any:
         return x
@@ -153,15 +132,6 @@ class Plot(Component):
             dtype = "altair" if is_altair else "plotly"
             out_y = y.to_json()
         return PlotData(**{"type": dtype, "plot": out_y})
-
-    def style(self, container: bool | None = None):
-        """
-        This method is deprecated. Please set these arguments in the constructor instead.
-        """
-        warn_style_method_deprecation()
-        if container is not None:
-            self.container = container
-        return self
 
 
 class AltairPlot: