+++ js/_website/src/lib/components/Header.svelte
@@ -40,8 +40,9 @@
 		<a class="thin-link flex items-center gap-3" href="/guides"
 			><span>💡</span> <span>Guides</span></a
 		>
-		<a class="thin-link flex items-center gap-3" href="/demos"
-			><span>🎢</span> <span>Demos</span></a
+		<a class="thin-link flex items-center gap-3" href="/playground"
+			><span>🎢</span>
+			<span>Playground<sup class="text-orange-500">NEW</sup></span></a
 		>
 		<div
 			on:mouseenter={() => (show_help_menu = true)}