+++ js/preview/src/svelte-internal.ts
@@ -1,1 +1,2 @@
+//@ts-ignore
 export * from "svelte/internal";